completely done using binary search -- error solved

+++ MySubmission/LeetCode/278.first-bad-version.cpp
@@ -11,13 +11,13 @@
 class Solution {
 public:
     int firstBadVersion(int n) {
-        int l = 1, r = n;
+        long long l = 1, r = n;
         while(l<r){
-            int mid = (l+r)/2;
+            long long mid = (l+r)/2;
             if(isBadVersion(mid)) r = mid;
             else l = mid+1;
         }
-        return l;
+        return (int)l;
     }
 };
 // @lc code=end
